move add_grating_couplers into gdsfactory.components.add_grating_couplers
