+++ .gitignore
@@ -7,8 +7,9 @@
 *.hdf5
 *.stl
 *.pkl
+test_mask_metadata.csv
+test_mask_metadata.yml
 .pyre_configuration
-gdsfactory/klayout/tech/layers.py
 poetry.lock
 mask.csv
 mask.yml
