+++ gdsfactory/containers.py
@@ -1,11 +1,11 @@
 import sys
 
-from gdsfactory.add_grating_couplers import (
+from gdsfactory.add_padding import add_padding_container
+from gdsfactory.add_termination import add_termination
+from gdsfactory.components.add_grating_couplers import (
     add_grating_couplers,
     add_grating_couplers_with_loopback_fiber_single,
 )
-from gdsfactory.add_padding import add_padding_container
-from gdsfactory.add_termination import add_termination
 from gdsfactory.components.array_component import array
 from gdsfactory.components.bend_port import bend_port
 from gdsfactory.components.cavity import cavity
