+++ gdsfactory/functions.py
@@ -126,7 +126,7 @@ def rotate(
         )
 
     component_new.add_ports(ref.ports)
-    component_new.info = component.info.copy()
+    component_new.copy_child_info(component)
     return component_new
 
 
@@ -154,7 +154,7 @@ def mirror(
     ref = component_new.add_ref(component)
     ref.mirror(p1=p1, p2=p2)
     component_new.add_ports(ref.ports)
-    component_new.info = component.info.copy()
+    component_new.copy_child_info(component)
     return component_new
 
 
@@ -221,7 +221,7 @@ def add_settings_label(
     settings: Optional[Strs] = None,
     ignore: Optional[Strs] = ("decorator",),
 ) -> Component:
-    """Add a settings label to a component.
+    """Add a settings label to a component. Use it as a decorator.
 
     Args:
         component: spec.
