+++ gdsfactory/labels/__init__.py
@@ -4,6 +4,7 @@
 from gdsfactory.labels import ehva, siepic, write_labels
 from gdsfactory.labels.add_label_yaml import add_label_yaml
 from gdsfactory.labels.ehva import DFT, Dft, add_label_ehva
+from gdsfactory.labels.merge_test_metadata import merge_test_metadata
 from gdsfactory.labels.siepic import add_fiber_array_siepic
 
 __all__ = [
@@ -15,4 +16,5 @@ __all__ = [
     "ehva",
     "siepic",
     "write_labels",
+    "merge_test_metadata",
 ]
